+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- Added `Text` Atom to handle attributed accessibility labels/hints in SwiftUI versions < iOS 15
+- Added `AttributedText` Atom to handle `NSAttributedString` in SwiftUI versions < iOS 15
+- Added `alignment: Alignment` parmeter to `SecondaryButtonStyle`
+- Fixed `EditListView` layout in large text mode
+- Fixed `IconButtonView` image height that was causing issues in large text mode
 
 ## [1.2.0] - 2023-03-02Z
 - Removed conflicting animations from donut chart